modify project directory
