+++ tests/test_signup.py
@@ -1,32 +1,11 @@
 import unittest
-from account import Account
-from signup import Signup
+from tests.basetest import BaseTestCase
 
 
-class TestUser(unittest.TestCase):
+class TestUser(BaseTestCase):
     """
     This class holds a setup method that creates the demo data to be used to test the program
     """
-
-    def setUp(self):
-        """
-        This method runs before each test by creating object sample data
-        :return:
-        """
-        self.user_signup = Signup("Calvin", "Peter", "calvinpete", "tinkacalvin@gmail.com", "<!Mario8Funke/>", "26",
-                                  "male")
-        self.user1_signup = Signup("Mark", "Ronald", "Mark", "markronald.com", "marky", "-6", "male")
-        self.user2_signup = Signup("Lonah", "Jemah", "Jemah", "Jemah8lo@hotmail.com", "j9W>olgm", "hjh", "male")
-        self.user3_signup = Signup("King", "David", "King David", "davidking@gmail.com", "psaLms198?", "124", "male")
-        self.user4_signup = Signup("Harvey", "Moses", "M H", "MH@gmail.com", "suQ3?kdright", "34", "male")
-        self.user_account = Account()
-
-    def test_creation(self):
-        """
-        This tests instance of a class
-        :return:
-        """
-        self.assertIsInstance(self.user_signup, Signup)
 
     def test_full_name(self):
         """
@@ -41,6 +20,8 @@ class TestUser(unittest.TestCase):
         :return:
         """
         self.assertEqual(self.user_signup.validate_password(), True)
+
+    def test_invalid_password(self):
         with self.assertRaises(ValueError):
             self.user1_signup.validate_password()
 
@@ -78,9 +59,6 @@ class TestUser(unittest.TestCase):
             self.user1_signup.validate_age()
         with self.assertRaises(ValueError):
             self.user2_signup.validate_age()
-
-    def tearDown(self):
-        pass
 
 
 if __name__ == "__main__":
